+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
+### Added
+
+- `OptimizationResults.to_xyz()` to convert the trajectory to an `xyz` multi-structure format.
+
 ## [0.10.1] - 2024-07-10
 
 ### Changed